+++ pom.xml
@@ -3,7 +3,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>fr.blossom.sample</groupId>
+  <groupId>com.blossom-project.sample</groupId>
   <artifactId>blossom-samples</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <packaging>pom</packaging>
