+++ sample-complete/src/main/java/com/blossom_project/sample/PouetController.java
@@ -1,0 +1,15 @@
+package com.blossom_project.sample;
+
+///**
+// * Created by maelg on 05/05/2017.
+// */
+//@BlossomController
+//@RequestMapping("/test/pouet")
+//@OpenedMenu("pouet")
+//public class PouetController {
+//
+//  @GetMapping
+//  public String pouet() {
+//    return "pouet";
+//  }
+//}
