--- sample-complete/src/main/java/fr/blossom/sample/Application.java
@@ -1,15 +1,0 @@
-
-package fr.blossom.sample;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-
-  public final static void main(String... args) {
-    SpringApplication.run(Application.class, args);
-  }
-
-}
