--- sample-complete/src/main/java/fr/blossom/sample/PouetController.java
@@ -1,15 +1,0 @@
-package fr.blossom.sample;
-
-///**
-// * Created by maelg on 05/05/2017.
-// */
-//@BlossomController
-//@RequestMapping("/test/pouet")
-//@OpenedMenu("pouet")
-//public class PouetController {
-//
-//  @GetMapping
-//  public String pouet() {
-//    return "pouet";
-//  }
-//}
