+++ sample-file-manager/pom.xml
@@ -4,7 +4,7 @@
   xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>fr.blossom.sample</groupId>
+  <groupId>com.blossom-project.sample</groupId>
   <artifactId>sample-file-manager</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <name>sample-file-manager</name>
@@ -21,7 +21,7 @@
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>fr.blossom</groupId>
+        <groupId>com.blossom-project</groupId>
         <artifactId>blossom-parent</artifactId>
         <version>${blossom.version}</version>
         <type>pom</type>
@@ -31,12 +31,12 @@
   </dependencyManagement>
   <dependencies>
     <dependency>
-      <groupId>fr.blossom</groupId>
+      <groupId>com.blossom-project</groupId>
       <artifactId>blossom-starter-ui-web</artifactId>
       <version>${blossom.version}</version>
     </dependency>
     <dependency>
-      <groupId>fr.blossom</groupId>
+      <groupId>com.blossom-project</groupId>
       <artifactId>blossom-starter-filemanager</artifactId>
       <version>${blossom.version}</version>
     </dependency>
