+++ sample-file-manager/src/main/java/com/blossom_project/sample/Application.java
@@ -1,0 +1,15 @@
+
+package com.blossom_project.sample;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+
+  public final static void main(String... args) {
+    SpringApplication.run(Application.class, args);
+  }
+
+}
