+++ sample-kotlin-java/pom.xml
@@ -4,7 +4,7 @@
   xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>fr.blossom.sample</groupId>
+  <groupId>com.blossom-project.sample</groupId>
   <artifactId>sample-kotlin-java</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <name>sample-kotlin-java</name>
@@ -22,7 +22,7 @@
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>fr.blossom</groupId>
+        <groupId>com.blossom-project</groupId>
         <artifactId>blossom-parent</artifactId>
         <version>${blossom.version}</version>
         <type>pom</type>
@@ -32,7 +32,7 @@
   </dependencyManagement>
   <dependencies>
     <dependency>
-      <groupId>fr.blossom</groupId>
+      <groupId>com.blossom-project</groupId>
       <artifactId>blossom-starter-ui-web</artifactId>
       <version>${blossom.version}</version>
     </dependency>
