+++ sample-kotlin-java/src/main/kotlin/com/blossom_project/sample/Application.kt
@@ -1,0 +1,11 @@
+package com.blossom_project.sample
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication
+open class Application
+
+fun main(args: Array<String>) {
+  SpringApplication.run(Application::class.java, *args)
+}
