--- sample-kotlin/src/main/kotlin/fr/blossom/sample/Application.kt
@@ -1,11 +1,0 @@
-package fr.blossom.sample
-
-import org.springframework.boot.SpringApplication
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-@SpringBootApplication
-open class Application
-
-fun main(args: Array<String>) {
-  SpringApplication.run(Application::class.java, *args)
-}
