+++ sample-war-packaging/pom.xml
@@ -4,7 +4,7 @@
   xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>fr.blossom.sample</groupId>
+  <groupId>com.blossom-project.sample</groupId>
   <artifactId>sample-war-packaging</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <packaging>war</packaging>
@@ -16,14 +16,14 @@
     <java.version>1.8</java.version>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
-    <start.class>fr.blossom.sample.Application</start.class>
+    <start.class>com.blossom-project.sample.Application</start.class>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <springboot.version>2.0.0.RC1</springboot.version>
   </properties>
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>fr.blossom</groupId>
+        <groupId>com.blossom-project</groupId>
         <artifactId>blossom-parent</artifactId>
         <version>${blossom.version}</version>
         <type>pom</type>
@@ -33,7 +33,7 @@
   </dependencyManagement>
   <dependencies>
     <dependency>
-      <groupId>fr.blossom</groupId>
+      <groupId>com.blossom-project</groupId>
       <artifactId>blossom-starter-ui-web</artifactId>
       <version>${blossom.version}</version>
     </dependency>
