+++ sample-war-packaging/src/main/java/com/blossom_project/sample/Application.java
@@ -1,0 +1,23 @@
+
+package com.blossom_project.sample;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+public class Application
+  extends SpringBootServletInitializer {
+
+
+  @Override
+  protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+    return application.sources(Application.class);
+  }
+
+  public final static void main(String... args) {
+    SpringApplication.run(Application.class, args);
+  }
+
+}
